huobi: refactor to base class
